Move QtMain.cpp and PCMain.cpp to more sensible locations. Native isn't standalone anymore...
